Add script for get it working on Windows 10
Under Windows 10 the ITC hardware does not work out of the box with
64bit Igor Pro.

The reason is that ASLR, see
https://de.wikipedia.org/wiki/Address_Space_Layout_Randomization, is
enabled by default and as it turns out breaks the ITCMM.dll in this
case.

Add a script to disable ASLR for Igor Pro 7 and 8. This assumes a
installation in the default folder of Igor Pro.

See
https://docs.microsoft.com/en-us/windows/security/threat-protection/windows-defender-exploit-guard/enable-exploit-protection#powershell
for some background on the powershell code.