ITCOpenDevice2: Make it more robust on errors
We have one report from a user (Tim Jarsky) that sometimes the device
open/close state of the XOP is not the same as the one from the ITC
library.

This can for example happen if the ITCDLL::ITC_InitDevice method throws.
We now remove the device again in this case from our internal list.

And we also postpone the return of the opened device ID after having
done all initialization.