MakeDestWave: Fix logic for /FREE flag
We only need to check calledFromFunction. No matter if we are running in a
thread (tp != NULL) or from main (tp == NULL) free waves can always be
used as long as we are running in functions.

We also use the more specialized error flag
FREE_FLAG_ALLOWED_IN_FUNCTIONS_ONLY.