MakeDestWave: Fix it again
The fix from 3993f51c (MakeDestWave: Fix free wave checking, 2022-06-10)
which fixed ea7b5f14 (MakeDestWave: Fix logic for /FREE flag, 2022-05-26)
was still not correct.

The calledFromMacro member can be set if a macro is running during a
preemptive thread.

So let's rework it again and take that into account as well.

The code here is correct with the current buggy Igor Pro behaviour and
hopefully also with future fixed version of Igor Pro.