CloseDeviceLowLevel: Add helper and make it more robust
We have occasionally problems running the MIES hardware tests [1] with
ITC1600 hardware. The reported problem [2] revolves around the fact that a
device can not be closed anymore, as closing it always returns the error

	Error in call to ITC_ConfigDevice.
	Reported error: DSP timeout (0x80303001)

We also already know that by restarting Igor Pro we can use that device
again.

So let's work around that issue by allowing ITC_ConfigDevice
to fail. From the best of this author's knowledge the ITC_ConfigDevice
call is only used for disabling the LED on the frontpanel of the hardware
box.

[1]: https://github.com/AllenInstitute/MIES
[2]: https://github.com/AllenInstitute/ITCXOP2/issues/22