.gitlab-ci.yml: Optimize git operations for jobs
The download-xoptoolkit and symbol_creation_and_upload jobs don't need an
up-to-date git checkout. This is set with GIT_STRATEGY.

The other jobs should also see all tags therefore we need to set GIT_DEPTH
to 0. But due to [1] this does not always work, so we need to unshallow
if required. In [2] the trick with (call ) is explained as:

  To clear the ERRORLEVEL back to 0, run (call ), which does nothing except
  set the ERRORLEVEL to 0.

[1]: https://gitlab.com/gitlab-org/gitlab/-/issues/292470
[2]: https://ss64.com/nt/errorlevel.html